requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
python-dotenv
cohere
langchain
langchain-cohere
